Battle2k: Fix bug in self destruct logic
Self destruct would set the enemy as hidden, and then later
RemoveActionsForNonExistantBattlers() would remove it before
it completed, causing the next battle algorithm to start running
midway.

If the next algo was a defend or something with no target, it will
crash. If it's an attack, it can cause the battle not to finish.

Fix #1875