Make the criteria for clearing the map screen (a single 320x240x4 memset) more strict: It is always cleared except when a panorama is active.
The reason for this is another edge case in the previous commit: The lower layer star tile can be partially transparent and the upper layer tile below, too.
In this case there will be a smear because the background is not drawn. This is not fixable without additional book-keeping but a screen clear solves it.