Refactor Window_Message
* Reorganize update sequence
* Add debug logging to debug frame timing issues
* Simplify and streamline
* Add hacks to match RPG_RT open/close timing and intpreter
* Fix waits for printable and non-printable characters.

It appears that RPG_RT animates the message open and close *after*
the main loop. This causes a lot of complications for us, because much
code depends on whether the message window is visible or closing.

To workaround this difference we add close started and close finished
flags and wrap them.