Optimize TilemapLayer::DrawTile
Ensure the check against transparent tiles is done before DrawTileImpl()
is called.

For transparent maps, this produces a measurable speedup as we never pay
the cost of calling DrawTileImpl() function.

For most maps, lower layers are likely to be opaque while upper layers
are likely to be transparent. This change results in separate branches
for each tile layer, and therefore allows the branch predictor to
predict both patterns.