Added CheckAnimFlip function
The CheckAnimFlip function has been added which takes a Game_Battler
as a parameter. This puts the whole XOR-Operations in one place.
Moreover the invert_animation flag has been removed because the
changes made storing the state unnecessary. And it is sufficient
to check the first original target on a reflected animation.