States: Damage and recovery is rounded down instead of up
If an actor get damaged or recovered due to a state, the damage/
recovery value is rounded down instead of up. This means that the
damage/recovery value can even be zero if the damage/recovery
percentage is set to a low value. Moreover a bug got fixed in SP
damage/recovery: The SP damage/recovery percentage uses now the
maximum SP instead of the maximum HP for the calculation.