Enemy AI changes: Change weighting of priorities and add special check
The weighting of the priorities has been changed: The action with the
highest priority always gets a weight of 10 and actions with a
priority of (highest priority - 9) gets a weight of 1. The scaling
between is linear. This means if we have two actions, one with a
priority of 50 and one with a priority of 45, the first action gets
a 10/15 chance and the second action a 5/15 chance to get selected.
Moreover a special check has been added if the enemy uses a skill
which only heals states without affecting anything else: He actually
checks if there is an ally which is affected by at least one state
which get healed by the skill. If this is not the case, then the
action gets discarded.