When the cache is exhausted do not delete unreferenced assets that were used during the last 50ms (approx. 3 frames).
The current behaviour (flush everything unreferenced when cache is full) was added to workaround out-of-memory issues on systems with limited RAM but this direct unload makes some use cases like face rendering in the Battle scene with Gauge style really slow (when the cache is full the face is unloaded and reloaded from disk once every frame).

I hope that 50ms is a good compromise here.

Fix #2509