Refactor AudioDecoders to chrono.
Looks at some places a bit ugly because chrono uses by default integers internally, so direct second to ms casts are not possible here.
Precision for the Update function was increased to microseconds to match the precision required for MIDI.

Also fixed the timing for the MIDI decoders. mtime was updated too often.