Add "easyrpg_critical_hit_rate" attribute for skills
Skills can do critical hits now and the "easyrpg_critical_hit_rate"
determines the probability of a critical hit in percent. If set to a
negative value, then the critical hit rate of the battler is used.
If set to zero (default value) then the skill never does critical
hits.