Add feature namespace
A feature namespace has been added which can be used to check if the
RPG Maker 2000 battle system or RPG Maker 2003 battle system is used,
if the row feature is enabled and if text placeholders are enabled.
Many lengthy if-statements which check those feature settings have
been changed to use the methods from the feature namespace to improve
the readability of the code.