Panorama positioning depends on width and height of the panorama.
In some cases (not scrolling or looping) the position is reset after loading the panorama.
Here was the bug: Multiple ChangePanorama commands overwrote the reset indicator.

Added a async yield op to the event and made the panorama image important to ensure that the image is loaded.

Refactored Spriteset_Map to prevent deletion as the panorama change is otherwise a dangling reference.

Fix #2195