Allow leaving tiles that are out-of-bounds.
Fixes a stuck Player in Yume2kki when teleporting from Christmas World to Exhibition.
The teleport target is out-of-bounds (maybe because the map was resized later) preventing leaving the tile.

This is not fully bug compatible to RPG_RT. Assuming the Y-Coordinate is out-of-bounds: When moving
left or right the invalid Y will stay in RPG_RT preventing events from being triggered, but we wrap it
inbounds after the first move.

Fix #2716