Make the number of level digits dependent on the global maximum level
The number of level digits is now dependent on the global maximum
level setting instead of the maximum level of the actor. This fixes
a small optical issue with the level indent if you have actors in
the party which can reach level 100 or greater and actors which
can reach at most level 99.