Support premultiplied alpha for Tone blit.
Also refactor ToneBlit function structure.

Replaced ImageOpacity::Partial with Alpha_1Bit and Alpha_8Bit.
This allows further optimisations in Tone Blit to decide if premultiply is needed or not.

Benchmarks have shown that on x86 the difference is not measurable between multiply or not.
Maybe helps on other platforms with slower multiply.

Fix #2755