Fix Player parsing BattleTest argument incorrectly
This broke the BattleTest feature of the editor.
When a battle test happens the arguments are:
BattleTest ShowTitle Window 1 2 3 4 5 where 1-5 are the arguments.
The arguments are not next to it.

Another edge case working: "--battle-test 1 Window".
Here Window argument was "eaten" because battle-test reads either 1 or 4 arguments.
Now the code rewinds correctly and processes the Window argument.