SDL1: Fix D-Pad and improve handling of axis
Mostly for Wii but will help other ports.

The axis mapping can be configured now by providing a GetSdlAxis function or through EP_SDL_AXIS environment variable.
Without this mapping the input code is not usable.

Remapping buttons was already possible before.

This is a bit like the ControllerDB in SDL2.