WIP - Add support to variables in both CommandPlayBGM and CommandPlaySound
Ok, I followed @Ghabry's suggestions and tried to implement a function to format the sound Effects and Music if it has variables attached to it values.

I'm stuck in 2 problems, 1st one:
There's a huge delay between times where the same SFX is played. It does not happens in Maniacs RPG_RT.exe. Is there a rule that waits finishing a sfx before playing it again?

2nd Problem: Weird Edge Case.

the hard-coded version of my code works out of the box:

```cpp
bool Game_Interpreter::CommandPlaySound(lcf::rpg::EventCommand const& com) { // code 11550

	std::array<int, 3> params = { com.parameters[0], com.parameters[1], com.parameters[2] };

	if (Player::IsPatchManiac() && com.parameters.size()>3)
	{
		params[0] = ValueOrVariable((com.parameters[3] & 0xF0) >> 4, com.parameters[0]);
		params[1] = ValueOrVariable((com.parameters[3] & 0xF00) >> 8, com.parameters[1]);
		params[2] = ValueOrVariable((com.parameters[3] & 0xF000) >> 12, com.parameters[2]);
	};

	lcf::rpg::Sound sound;
	sound.name = ToString(com.string);
	sound.volume = params[0];
	sound.tempo = params[1];
	sound.balance = params[2];
	Main_Data::game_system->SePlay(sound, true);
	return true;
}
```

But the dynamic version presented in this PR does not work as intended in some cases.

Here's the game I'm using to test it:
https://github.com/EasyRPG/liblcf/files/10188776/Project2.zip

the pitch shift work in most of the sound effects. But if I hold enter and there's a pitch shift in aaa.wav, It gives a very low value, near to zero.

Again, that's does not happens on the hardcoded snippet I wrote before.

Anyone can help me with those issues?