Add ExecuteCommand(com)
Minor Change.
It's a variation of ExecuteCommand(); that accepts `com` as a parameter.

Ghabry commented that this would make the code faster.
But, this also open possibilities to create a scripting language, since we can create `com` parameters from strings.

------------------

This implementation could also be used to attach code from dynRPG patches to new `com` stuff.
Having both @comments and `com` commands share the same code.