Add Drive Filesystem
Is a virtual filesystem that lists drive letters on e.g. Windows.
The filesystem is deactivated on all other platforms right now.

Invented a new filetype "Filesystem" for delegation of the drive FS to the native FS.
This is kinda a hack but I couldn't think of a better way to add this...