Fix flickering in Yume2kki on map 3D Underworld (ID 1884).
The map uses a MoveRoute with a jump and SetVehicleLocation for party movement in a tight loop which causes heavy flickering in our Player.

This fix does not appear to be completely correct as RPG_RT does not reset the jump flag here but the "damage" is reduced because SetVehicleLocation -1 cannot happen without patching the game.

Fix #3254

Co-Authored-By: Viet Dinh <54ckb0y789@gmail.com>