New Command 2059 - EasyRPG Zoom
Zooms in and Out the screen.

`The syntax always comes in pairs. The first parameter of each pair indicates whether you are using a direct value or a variable/indirect variable, through ValueOrVariable().`

------------------------

#### Syntax (TPC):
```js
@raw 2059,
      "",          // Not Used.
      0, 0,        // Scale value in percentage (Default is 100%).
      0, 1,        // Origin Type (0 = absolute pixel || 1 =percentage of screen area )
      0, 50,        // Origin X
      0, 50,        // Origin Y
```

#### Example:
```js
@raw 2059, "", 1, 20, 0, 0, 1, 21, 1, 22 // size is var20, absolute pixel, originX is var21, originY is var22
```

-----------------------------------

This command differs from maniacs patch zoom, which is executed per picture layer, and it has a interpolation between zoom states and a "wait for completion" check.