Fix StringIsAscii by explicitly setting the type to "signed char"
Turns out "char" is not always signed. Instead it is "implementation defined".
In all tests it went unnoticed because x86 is "signed". On ARM it appears to be
"unsigned".

This is surprising. For int etc. it is specified to be signed.

Can we just burn the C standard? Who specifies this nonsense?

Fix #3401