Consider images with no alpha as Opaque in ComputeImageOpacity.
When using pixel format ABGR (prefered on Big Endian) instead of RGBA pixman makes the alpha pixels 0x00 instead of 0xFF.

Technically the alpha values are undefined as the image is opaque but our opacity function considered this image as full transparent (because A is 0 everywhere).

This triggered a fast path in the ToneBlit which does nothing when the image is full transparent and causes a "trippy trail" rendering glitch.

Additionally this fixes a long-standing bug in "Unterwegs in Düsterburg" on the Wii/U: The fog effect at the beginning only flickered between transparent and full opaque before.

Fix #3436