NEW COMMAND 2059 - Typing mode
Allow the player to Type some info inside the game.

It's headless implementation, only way to get a feedback that the player is in typing mode is by having a paralell process displaying the stringvariable that is being edited.

command is
```js
@cmd 2059,
"Default Name", // default value to be set over into the variable.
 [
0, 3,  // output stringvar mode and output stringvar id
0, 0, // default value as a stringvar mode and id
1, // allow multilines when hitting (ctrl + enter)
]

```